[Misc][Refactor] Separate router from FusedMoE class
